fix: NixOS module and test
The config/example.json is used as Brockman's configuration within the
NixOS test. However, it was configured with `"tls": true`, while the
charybdis IRC daemon started without TLS.

Since NixOS 21.05, one of isNormalUser or isSystemUser needs to be set.

services.mingetty is an alias for services.getty, producing a
deprecation warning.